+++ PluginsAndFeatures/azure-toolkit-for-intellij/src/com/microsoft/azure/toolkit/intellij/springcloud/creation/SpringCloudAppCreationDialog.form
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<form xmlns="http://www.intellij.com/uidesigner/form/" version="1" bind-to-class="com.microsoft.azure.toolkit.intellij.springcloud.component.SpringCloudAppCreationDialog">
+<form xmlns="http://www.intellij.com/uidesigner/form/" version="1" bind-to-class="com.microsoft.azure.toolkit.intellij.springcloud.creation.SpringCloudAppCreationDialog">
   <grid id="27dc6" binding="panel" layout-manager="GridLayoutManager" row-count="2" column-count="1" same-size-horizontally="false" same-size-vertically="false" hgap="-1" vgap="-1">
     <margin top="0" left="0" bottom="0" right="0"/>
     <constraints>
@@ -8,12 +8,12 @@
     <properties/>
     <border type="none"/>
     <children>
-      <nested-form id="441f4" form-file="com/microsoft/azure/toolkit/intellij/springcloud/component/SpringCloudAppInfoBasicPanel.form" binding="basicForm" custom-create="true">
+      <nested-form id="441f4" form-file="com/microsoft/azure/toolkit/intellij/springcloud/creation/SpringCloudAppInfoBasicPanel.form" binding="basicForm" custom-create="true">
         <constraints>
           <grid row="0" column="0" row-span="1" col-span="1" vsize-policy="3" hsize-policy="3" anchor="1" fill="1" indent="0" use-parent-layout="false"/>
         </constraints>
       </nested-form>
-      <nested-form id="70345" form-file="com/microsoft/azure/toolkit/intellij/springcloud/component/SpringCloudAppInfoAdvancedPanel.form" binding="advancedForm" custom-create="true">
+      <nested-form id="70345" form-file="com/microsoft/azure/toolkit/intellij/springcloud/creation/SpringCloudAppInfoAdvancedPanel.form" binding="advancedForm" custom-create="true">
         <constraints>
           <grid row="1" column="0" row-span="1" col-span="1" vsize-policy="3" hsize-policy="3" anchor="1" fill="1" indent="0" use-parent-layout="false"/>
         </constraints>
